scope informational component to ama appeals
